move local storage logic from App component to separate hook
