+++ src/App.vue
@@ -26,6 +26,7 @@ import GameBoard from "./components/GameBoard.vue";
 import VButton from "./components/VButton.vue";
 import RulesModal from "./components/RulesModal.vue";
 import { ref, onMounted } from "vue";
+import useLocalStore from "./hooks/useLocalStore";
 export default {
   components: { VHeader, VButton, GameBoard, ScoreBoard, RulesModal },
   setup(){
@@ -53,35 +54,9 @@ export default {
     }
     // game score state
     const score = ref(0);
-    const localStorageKey = "rpsls-score";
+    const { loadScore, saveScore } = useLocalStore();
     
-    const isLocalStorageAvailable = () => {
-      const name = "localStorageTest-RPSLS";
-      try {
-        localStorage.setItem(name,name);
-        localStorage.removeItem(name);
-        return true;
-      } catch(e) {
-        console.warn("local storage not available");
-        return false;
-      }
-    }
-    const saveScore = () => {
-      if(isLocalStorageAvailable()){
-        localStorage.setItem(localStorageKey,JSON.stringify(score.value));
-      }
-    };
-
-    const loadScore = () => {
-      if(isLocalStorageAvailable()){
-        let savedScore = localStorage.getItem(localStorageKey);
-        if(savedScore){
-          score.value = JSON.parse(savedScore);
-        } 
-      }
-
-    };
-  onMounted(loadScore);
+  onMounted(() => loadScore(score));
 
     const incrementScore = () => {
       score.value++;
@@ -139,7 +114,7 @@ export default {
             setStage(gameStages.lost);
             if(score.value > 0) decrementScore();
           }
-          saveScore();
+          saveScore(score.value);
         }
 
       const selectPlayerFigure = figure => {
